fix numba reflected list deprecation warning

+++ fteikpy/_solver.py
@@ -23,7 +23,7 @@ class Eikonal2D(BaseGrid2D):
         super().__init__(
             grid=grid,
             gridsize=gridsize,
-            origin=origin if origin is not None else numpy.zeros(2),
+            origin=origin if origin is not None else numpy.zeros(2, dtype=numpy.float64),
         )
 
     def solve(self, sources, nsweep=2, return_gradient=False):
@@ -95,7 +95,7 @@ class Eikonal3D(BaseGrid3D):
         super().__init__(
             grid=grid,
             gridsize=gridsize,
-            origin=origin if origin is not None else numpy.zeros(3),
+            origin=origin if origin is not None else numpy.zeros(3, dtype=numpy.float64),
         )
 
     def solve(self, sources, nsweep=2, return_gradient=False):
